requirements.txt
streamlit
numpy
opencv-python
tensorflow
pillow
